requirements.txt
fastapi[all]
uvicorn
minio
pyyaml
requests
Fancy_term
